Tightly integrate ExponentialGrowth estimator with bin-packing node allocator
It is convenient to have those two coupled so we can make use of workers
memory pool information within ExponentialGrowthPartitionMemoryEstimator
